+++ src/compose/compose-node.ts
@@ -2,7 +2,7 @@ import type { Directives } from '../doc/directives.js'
 import { Alias } from '../nodes/Alias.js'
 import type { ParsedNode } from '../nodes/Node.js'
 import type { ParseOptions } from '../options.js'
-import type { FlowScalar, Token } from '../parse/tokens.js'
+import type { FlowScalar, Token } from '../parse/cst.js'
 import type { Schema } from '../schema/Schema.js'
 import { composeCollection } from './compose-collection.js'
 import { composeScalar } from './compose-scalar.js'
