+++ src/compose/util-contains-newline.ts
@@ -1,4 +1,4 @@
-import type { Token } from '../parse/tokens.js'
+import type { Token } from '../parse/cst.js'
 
 export function containsNewline(key: Token | null | undefined) {
   if (!key) return null
