+++ src/parse/parser.ts
@@ -1,4 +1,3 @@
-import { Lexer } from './lexer.js'
 import {
   SourceToken,
   Token,
@@ -12,7 +11,8 @@ import {
   prettyToken,
   tokenType,
   TokenType
-} from './tokens.js'
+} from './cst.js'
+import { Lexer } from './lexer.js'
 
 function includesToken(list: SourceToken[], type: SourceToken['type']) {
   for (let i = 0; i < list.length; ++i) if (list[i].type === type) return true
